rename: 버튼 형식의 공통 컴포넌트 @/components/common/button으로 이동
